mixin refactor, with test function
